provide non static methods
Move all logic to instance methods instead of static methods.
The static interface remains the same, but its implementation
uses an instance (object) internally.

This allows the user to choose between static interface and
a service object. The latter may especially be useful if
a service container (DI container) is used and the services
are injected.